complete admin-side U2F key registration

+++ src/lib/src/Modules/LoginGuard/Lib/TwoFactor/Provider/U2F.php
@@ -10,9 +10,14 @@ class U2F extends BaseProvider {
 
 	const SLUG = 'u2f';
 
+	/**
+	 * @var RegisterRequest
+	 */
+	private $oWorkingRegistration;
+
 	public function setup() {
 		add_action( 'admin_enqueue_scripts', function ( $sHook ) {
-			if ( in_array( $sHook, [ 'profile.php', ] ) ) {
+			if ( in_array( $sHook, [ 'profile.php', ] ) && Services::WpUsers()->isUserLoggedIn() ) {
 				$this->enqueueAdminU2f();
 			}
 		} );
@@ -29,26 +34,22 @@ class U2F extends BaseProvider {
 			$aDeps[] = $this->getCon()->prefix( $sScript );
 		}
 
+		$oUser = Services::WpUsers()->getCurrentWpUser();
+		$bValidated = $this->hasValidatedProfile( $oUser );
+		error_log( var_export( $bValidated, true ) );
 		try {
-			/** @var RegisterRequest $oReq */
-			$aParts = wp_parse_url( Services::WpGeneral()->getHomeUrl() );
-			$sAppId = sprintf( 'https://%s%s', $aParts[ 'host' ], empty( $aParts[ 'port' ] ) ? '' : ':'.$aParts[ 'port' ] );
-			list( $oReq, $signatures ) = ( new \u2flib_server\U2F( $sAppId ) )->getRegisterData( [] );
-
 			wp_localize_script(
 				$this->getCon()->prefix( 'shield-u2f-admin' ),
 				'icwp_wpsf_vars_u2f',
 				[
-					'raw_request' => (array)$oReq,
-					'request'     => [
-						'version'   => $oReq->version,
-						'challenge' => $oReq->challenge
+					'registration' => $bValidated ? [] : json_decode( $this->resetSecret( $oUser ) ),
+					'flags'        => [
+						'is_validated' => $bValidated
 					],
-					'app_id'      => $oReq->appId,
-					'strings'     => [
+					'strings'      => [
 						'not_supported' => __( 'U2F Security Key registration is not supported in this browser', 'wp-simple-firewall' ),
-						'failed'       => __( 'Key registration failed.', 'wp-simple-firewall' )
-										  .' '.__( 'Please retry or refresh the page.', 'wp-simple-firewall' ),
+						'failed'        => __( 'Key registration failed.', 'wp-simple-firewall' )
+										   .' '.__( 'Please retry or refresh the page.', 'wp-simple-firewall' ),
 						'do_save'       => __( 'Key registration was successful.', 'wp-simple-firewall' )
 										   .' '.__( 'Please now save your profile settings.', 'wp-simple-firewall' )
 					]
@@ -60,15 +61,56 @@ class U2F extends BaseProvider {
 	}
 
 	/**
+	 * @param \WP_User $oUser
+	 * @return RegisterRequest
+	 * @throws \u2flib_server\Error
+	 */
+	private function getU2fRegistration( \WP_User $oUser ) {
+		if ( !isset( $this->oWorkingRegistration ) ) {
+			// TODO: support multiple signatures
+			list( $this->oWorkingRegistration, $signatures ) = ( new \u2flib_server\U2F( $this->getU2fAppID() ) )->getRegisterData( [] );
+		}
+		return $this->oWorkingRegistration;
+	}
+
+	/**
+	 * @return string
+	 */
+	private function getU2fAppID() {
+		$aParts = wp_parse_url( Services::WpGeneral()->getHomeUrl() );
+		return sprintf( 'https://%s%s', $aParts[ 'host' ], empty( $aParts[ 'port' ] ) ? '' : ':'.$aParts[ 'port' ] );
+	}
+
+	/**
+	 * @param \WP_User $oUser
+	 * @return string
+	 */
+	protected function genNewSecret( \WP_User $oUser ) {
+		try {
+			return json_encode( (object)$this->getU2fRegistration( $oUser ) );
+		}
+		catch ( \Exception $oE ) {
+			return '';
+		}
+	}
+
+	/**
 	 * @inheritDoc
 	 */
 	public function renderUserProfileOptions( \WP_User $oUser ) {
+
+		$bValidated = $this->hasValidatedProfile( $oUser );
+
 		$aData = [
 			'strings' => [
-				'title'                 => __( 'U2F', 'wp-simple-firewall' ),
-				'button_reg_key'        => __( 'Register A New U2F Security Key', 'wp-simple-firewall' ),
-				'prompt'                => __( 'Click To Start U2F Security Registration', 'wp-simple-firewall' ),
-			]
+				'title'           => __( 'U2F', 'wp-simple-firewall' ),
+				'button_reg_key'  => __( 'Register A New U2F Security Key', 'wp-simple-firewall' ),
+				'prompt'          => __( 'Click To Start U2F Security Registration.', 'wp-simple-firewall' ),
+				'check_to_delete' => __( 'Check the box to delete your existing U2F key registration.', 'wp-simple-firewall' ),
+			],
+			'flags'   => [
+				'is_validated' => $bValidated
+			],
 		];
 
 		return $this->getMod()
@@ -87,45 +129,50 @@ class U2F extends BaseProvider {
 	}
 
 	/**
-	 * @return array
+	 * @inheritDoc
 	 */
-	public function getFormField() {
-		return [
-			'name'        => $this->getLoginFormParameter(),
-			'type'        => 'text',
-			'value'       => '',
-			'placeholder' => __( 'Please use your Backup Code to login.', 'wp-simple-firewall' ),
-			'text'        => __( 'Login Backup Code', 'wp-simple-firewall' ),
-			'help_link'   => '',
-		];
+	public function handleUserProfileSubmit( \WP_User $oUser ) {
+		$bError = false;
+		$sMsg = null;
+
+		$sU2fResponse = Services::Request()->post( 'icwp_new_u2f_response' );
+		if ( !$this->hasValidatedProfile( $oUser ) && !empty( $sU2fResponse ) ) {
+			try {
+				$aReg = json_decode( $this->getSecret( $oUser ), true );
+				$oRegistration = ( new \u2flib_server\U2F( $this->getU2fAppID() ) )->doRegister(
+					new RegisterRequest( $aReg[ 'challenge' ], $aReg[ 'appId' ] ),
+					json_decode( $sU2fResponse )
+				);
+
+				$this->setSecret( $oUser, json_encode( get_object_vars( $oRegistration ) ) )
+					 ->setProfileValidated( $oUser );
+
+				$sMsg = __( 'U2F Key has been registered successfully on your profile.', 'wp-simple-firewall' );
+			}
+			catch ( \Exception $oE ) {
+				$bError = true;
+				$sMsg = sprintf( __( 'U2F Key registration failed with the following error: %s', 'wp-simple-firewall' ),
+					$oE->getMessage() );
+			}
+		}
+		elseif ( Services::Request()->post( 'icwp_u2f_key_delete' ) === 'Y' ) {
+			$this->deleteSecret( $oUser )
+				 ->setProfileValidated( $oUser, false );
+			$sMsg = __( 'Registered U2F Key has been removed from your profile.', 'wp-simple-firewall' );
+		}
+
+		if ( !empty( $sMsg ) ) {
+			$this->getMod()->setFlashAdminNotice( $sMsg, $bError );
+		}
 	}
 
 	/**
-	 * @param \WP_User $oUser
-	 * @return bool
-	 */
-	public function hasValidatedProfile( $oUser ) {
-		return $this->hasValidSecret( $oUser );
-	}
-
-	/**
-	 * @param \WP_User $oUser
-	 * @return $this
-	 */
-	public function postSuccessActions( \WP_User $oUser ) {
-		$this->deleteSecret( $oUser );
-		$this->sendBackupCodeUsedEmail( $oUser );
-		return $this;
-	}
-
-	/**
-	 * Backup Code are 1-time only and if you have MFA, then we need to remove all the other tracking factors
 	 * @param \WP_User $oUser
 	 * @param string   $sOtpCode
 	 * @return bool
 	 */
 	protected function processOtp( $oUser, $sOtpCode ) {
-		return $this->validateBackupCode( $oUser, $sOtpCode );
+		return $this->validateU2F( $oUser, $sOtpCode );
 	}
 
 	/**
@@ -133,8 +180,8 @@ class U2F extends BaseProvider {
 	 * @param string   $sOtpCode
 	 * @return bool
 	 */
-	private function validateBackupCode( $oUser, $sOtpCode ) {
-		return wp_check_password( str_replace( '-', '', $sOtpCode ), $this->getSecret( $oUser ) );
+	private function validateU2F( $oUser, $sOtpCode ) {
+		return false; // TODO
 	}
 
 	/**
@@ -143,22 +190,14 @@ class U2F extends BaseProvider {
 	 */
 	protected function auditLogin( $oUser, $bIsSuccess ) {
 		$this->getCon()->fireEvent(
-			$bIsSuccess ? '2fa_backupcode_verified' : '2fa_backupcode_fail',
+			$bIsSuccess ? '2fa_u2f_verified' : '2fa_u2f_fail',
 			[
 				'audit' => [
 					'user_login' => $oUser->user_login,
-					'method'     => 'Backup Code',
+					'method'     => 'U2F',
 				]
 			]
 		);
-	}
-
-	/**
-	 * @param \WP_User $oUser
-	 * @return string
-	 */
-	protected function genNewSecret( \WP_User $oUser ) {
-		return wp_generate_password( 25, false );
 	}
 
 	/**
@@ -168,38 +207,5 @@ class U2F extends BaseProvider {
 		/** @var LoginGuard\Options $oOpts */
 		$oOpts = $this->getOptions();
 		return $oOpts->isEnabledU2F();
-	}
-
-	/**
-	 * @param \WP_User $oUser
-	 * @param string   $sNewSecret
-	 * @return $this
-	 */
-	protected function setSecret( $oUser, $sNewSecret ) {
-		parent::setSecret( $oUser, wp_hash_password( $sNewSecret ) );
-		return $this;
-	}
-
-	/**
-	 * @param \WP_User $oUser
-	 */
-	private function sendBackupCodeUsedEmail( $oUser ) {
-		$aEmailContent = [
-			__( 'This is a quick notice to inform you that your Backup Login code was just used.', 'wp-simple-firewall' ),
-			__( "Your WordPress account had only 1 backup login code.", 'wp-simple-firewall' )
-			.' '.__( "You must go to your profile and regenerate a new code if you want to use this method again.", 'wp-simple-firewall' ),
-			'',
-			sprintf( '<strong>%s</strong>', __( 'Login Details', 'wp-simple-firewall' ) ),
-			sprintf( '%s: %s', __( 'URL', 'wp-simple-firewall' ), Services::WpGeneral()->getHomeUrl() ),
-			sprintf( '%s: %s', __( 'Username', 'wp-simple-firewall' ), $oUser->user_login ),
-			sprintf( '%s: %s', __( 'IP Address', 'wp-simple-firewall' ), Services::IP()->getRequestIp() ),
-			'',
-			__( 'Thank You.', 'wp-simple-firewall' ),
-		];
-
-		$sTitle = sprintf( __( "Notice: %s", 'wp-simple-firewall' ), __( "Backup Login Code Just Used", 'wp-simple-firewall' ) );
-		$this->getMod()
-			 ->getEmailProcessor()
-			 ->sendEmailWithWrap( $oUser->user_email, $sTitle, $aEmailContent );
 	}
 }
